requirements.txt
# Athena Server v2 - Dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# HTTP client
httpx==0.27.2

# Database
psycopg2-binary==2.9.9

# AI SDKs
openai==1.54.0
anthropic==0.39.0

# Google APIs
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
